requirements.txt
asgiref==3.5.2
astroid==2.11.6
dill==0.3.5.1
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
flake8==4.0.1
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.14.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.1
typing_extensions==4.3.0
whitenoise==6.2.0
wrapt==1.14.1
